@@ -3,11 +3,11 @@ name = "nestos-installer"
 repository = "https://gitee.com/openeuler/nestos-installer"
 license = "Apache-2.0"
 edition = "2021"
-rust-version = "1.58.0"
+rust-version = "1.66.0"
 exclude = ["/.cci.jenkinsfile", "/.github", "/.gitignore", "/Dockerfile"]
 authors = [ "Benjamin Gilbert <bgilbert@redhat.com>" ]
 description = "Installer for NestOS"
-version = "0.16.0"
+version = "0.17.0"
 
 [package.metadata.release]
 sign-commit = true
@@ -18,19 +18,20 @@ pre-release-commit-message = "cargo: nestos-installer release {{version}}"
 pre-release-hook = ["make", "docs", "clean"]
 tag-message = "nestos-installer v{{version}}"
 
-# See https://github.com/cgwalters/cargo-vendor-filterer
+# See https://github.com/coreos/cargo-vendor-filterer
 [package.metadata.vendor-filter]
-# Right now the vendor filter doesn't support wildcards, so use this as a stand-in.
-# xref https://github.com/cgwalters/cargo-vendor-filterer/issues/23
-# Note that architecture-specific crate dependencies (apart from wasm32) are uncommon
-# in the ecosystem, but this project actually does have s390x-specific dependencies (below).
+# cargo-vendor-filterer supports wildcards but requires some degree of LLVM
+# support for every matching platform, which isn't necessarily available
+# out of the box.  Use these as stand-ins.  Note that architecture-specific
+# crate dependencies (apart from wasm32) are uncommon in the ecosystem, but
+# this project actually does have s390x-specific dependencies (below).
 platforms = ["x86_64-unknown-linux-gnu", "s390x-unknown-linux-gnu"]
 all-features = true
 
 [features]
 # rdcore is only useful inside the initrd of a NestOS system
 rdcore = []
-docgen = ["clap_mangen"]
+docgen = ["dep:clap_mangen"]
 
 [lib]
 name = "libcoreinst"
@@ -51,7 +52,7 @@ debug = true
 
 [dependencies]
 anyhow = ">= 1.0.38, < 2"
-base64 = "0.13"
+base64 = ">= 0.21, < 0.22"
 bincode = "^1.3"
 bytes = ">= 1.0.1, < 2"
 byte-unit = ">= 3.1.0, < 5.0.0"
@@ -63,30 +64,30 @@ glob = "^0.3"
 # disable default-enabled cli in gptman 0.x
 gptman = { version = ">= 0.7, < 2", default-features = false }
 hex = "^0.4"
-ignition-config = "0.2"
+ignition-config = ">= 0.3, < 0.4"
 lazy_static = "^1.4"
 libc = "^0.2"
-nix = ">= 0.24, < 0.26"
-nmstate = "2.1.2"
+nix = ">= 0.24, < 0.27"
+nmstate = { version = ">= 2.2.3, < 3", default-features = false, features = ["gen_conf"] }
 openssl = "^0.10"
 pipe = ">= 0.3, < 0.5"
 regex = ">= 1.4, < 2"
 reqwest = { version = ">= 0.10, < 0.12", features = ["blocking"] }
 serde = { version = "^1.0", features = ["derive"] }
 serde_json = "^1.0"
-serde_with = ">= 1.9.4, < 2"
+serde_with = ">= 1.9.4, < 3"
 serde_yaml = ">= 0.8, < 0.10"
 tempfile = ">= 3.1, < 4"
-textwrap = { version = ">= 0.15.0, < 0.16.0", default-features = false }
+textwrap = { version = ">= 0.15.0, < 0.17.0", default-features = false }
 thiserror = "1.0"
 url = ">= 2.1, < 3.0"
 uuid = { version = ">= 0.8, < 2.0", features = ["v4"] }
 walkdir = "^2.3"
 xz2 = "^0.1"
-zstd = { version = ">= 0.10.0, < 0.12.0", features = ["pkg-config"] }
+zstd = { version = ">= 0.10.0, < 0.13.0", features = ["pkg-config"] }
 
 [target.'cfg(target_arch = "s390x")'.dependencies]
-mbrman = { version = ">= 0.3, < 0.5", default-features = false }
+mbrman = ">= 0.5, < 0.6"
 rand = ">= 0.7, < 0.9"
 
 [dev-dependencies]
