@@ -45,7 +45,7 @@ pub fn download(config: DownloadConfig) -> Result<()> {
             config.fetch_retries,
         )?)
     };
-    eprintln!("{}", location);
+    eprintln!("{location}");
 
     // walk sources
     let mut sources = location.sources()?;
@@ -67,7 +67,7 @@ pub fn download(config: DownloadConfig) -> Result<()> {
         let mut path = PathBuf::new();
         path.push(&config.directory);
         path.push(filename);
-        let sig_path = path.with_file_name(format!("{}.sig", filename));
+        let sig_path = path.with_file_name(format!("{filename}.sig"));
 
         // check existing image and signature; don't redownload if OK
         // If we decompressed last time, the call will fail because we can't
@@ -359,8 +359,7 @@ pub fn image_copy_default(
     let offset = match saved {
         Some(saved) if saved.is_saved() => {
             // copy MBR
-            dest.seek(SeekFrom::Start(0))
-                .context("seeking disk to MBR")?;
+            dest.rewind().context("seeking disk to MBR")?;
             dest.write_all(&first_mb[0..saved.get_sector_size() as usize])
                 .context("writing MBR")?;
 
@@ -381,7 +380,7 @@ pub fn image_copy_default(
     };
     // do the copy
     dest.seek(SeekFrom::Start(offset))
-        .with_context(|| format!("seeking disk to offset {}", offset))?;
+        .with_context(|| format!("seeking disk to offset {offset}"))?;
     dest.write_all(&first_mb[offset as usize..first_mb.len()])
         .context("writing first MiB of disk")?;
 
@@ -399,13 +398,13 @@ pub fn download_to_tempfile(url: &Url, retries: FetchRetries) -> Result<File> {
         &mut BufReader::with_capacity(BUFFER_SIZE, &mut resp),
         &mut writer,
     )
-    .with_context(|| format!("couldn't copy '{}'", url))?;
+    .with_context(|| format!("couldn't copy '{url}'"))?;
     writer
         .flush()
-        .with_context(|| format!("couldn't write '{}' to disk", url))?;
+        .with_context(|| format!("couldn't write '{url}' to disk"))?;
     drop(writer);
-    f.seek(SeekFrom::Start(0))
-        .with_context(|| format!("rewinding file for '{}'", url))?;
+    f.rewind()
+        .with_context(|| format!("rewinding file for '{url}'"))?;
 
     Ok(f)
 }
@@ -426,7 +425,7 @@ struct ProgressReader<'a, R: Read> {
 impl<'a, R: Read> ProgressReader<'a, R> {
     fn new(source: R, length: Option<u64>, artifact_type: &'a str) -> Self {
         let tty = isatty(stderr().as_raw_fd()).unwrap_or_else(|e| {
-            eprintln!("checking if stderr is a TTY: {}", e);
+            eprintln!("checking if stderr is a TTY: {e}");
             false
         });
         // disable percentage reporting for zero-length files to avoid
@@ -647,7 +646,7 @@ mod tests {
         let saved = SavedPartitions::new_from_file(
             &mut dest,
             512,
-            &vec![PartitionFilter::Label(glob::Pattern::new("*").unwrap())],
+            &[PartitionFilter::Label(glob::Pattern::new("*").unwrap())],
         )
         .unwrap();
         assert!(saved.is_saved());
@@ -655,7 +654,7 @@ mod tests {
         let precious = "hello world";
         dest.seek(SeekFrom::Start(offset)).unwrap();
         dest.write_all(precious.as_bytes()).unwrap();
-        dest.seek(SeekFrom::Start(0)).unwrap();
+        dest.rewind().unwrap();
 
         let err = write_image(
             &mut FileLocation::new(source_path.to_str().unwrap())
@@ -672,9 +671,8 @@ mod tests {
         )
         .unwrap_err();
         assert!(
-            format!("{:#}", err).contains("collision with partition"),
-            "incorrect error: {:#}",
-            err
+            format!("{err:#}").contains("collision with partition"),
+            "incorrect error: {err:#}"
         );
 
         dest.seek(SeekFrom::Start(offset)).unwrap();
@@ -700,7 +698,7 @@ mod tests {
         source.seek(SeekFrom::Start(mb as u64)).unwrap();
         image_copy_default(&data[0..mb], &mut source, &mut dest, Path::new("/z"), None).unwrap();
         // compare
-        dest.seek(SeekFrom::Start(0)).unwrap();
+        dest.rewind().unwrap();
         let mut result = vec![0u8; len];
         dest.read_exact(&mut result).unwrap();
         assert_eq!(data, result);
@@ -711,7 +709,7 @@ mod tests {
         // gptman requires a fixed disk length
         dest.set_len(len as u64).unwrap();
         // create saved
-        let saved = SavedPartitions::new_from_file(&mut dest, 512, &vec![]).unwrap();
+        let saved = SavedPartitions::new_from_file(&mut dest, 512, &[]).unwrap();
         assert!(!saved.is_saved());
         // copy
         source.seek(SeekFrom::Start(mb as u64)).unwrap();
@@ -724,7 +722,7 @@ mod tests {
         )
         .unwrap();
         // compare
-        dest.seek(SeekFrom::Start(0)).unwrap();
+        dest.rewind().unwrap();
         let mut result = vec![0u8; len];
         dest.read_exact(&mut result).unwrap();
         assert_eq!(data, result);
@@ -744,7 +742,7 @@ mod tests {
         let saved = SavedPartitions::new_from_file(
             &mut dest,
             512,
-            &vec![PartitionFilter::Label(glob::Pattern::new("bovik").unwrap())],
+            &[PartitionFilter::Label(glob::Pattern::new("bovik").unwrap())],
         )
         .unwrap();
         assert!(saved.is_saved());
@@ -759,7 +757,7 @@ mod tests {
         )
         .unwrap();
         // compare
-        dest.seek(SeekFrom::Start(0)).unwrap();
+        dest.rewind().unwrap();
         let mut result = vec![0u8; len];
         dest.read_exact(&mut result).unwrap();
         assert_eq!(detect_formatted_sector_size(&result), NonZeroU32::new(512));
