@@ -38,11 +38,11 @@ const COREOS_ISO_FEATURES_PATH: &str = "NESTOS/FEATURES.JSO";
 #[derive(Default, Deserialize)]
 #[serde(default, rename_all = "kebab-case")]
 pub(super) struct OsFeatures {
-    /// Installer reads config files from /etc/coreos/installer.d
+    /// Installer reads config files from /etc/nestos/installer.d
     pub installer_config: bool,
     /// Directives supported in installer config files
     pub installer_config_directives: InstallerDirectives,
-    /// Live initrd reads NM keyfiles from /etc/coreos-firstboot-network
+    /// Live initrd reads NM keyfiles from /etc/nestos-firstboot-network
     pub live_initrd_network: bool,
 }
 
@@ -74,13 +74,13 @@ pub(super) struct LiveInitrd {
     /// The initrd for the live system
     initrd: Initrd,
     /// The Ignition config for the live system
-    live: Ignition,
+    live: Option<Ignition>,
     /// The Ignition config for the destination system
     dest: Option<Ignition>,
     /// User-supplied Ignition configs for the dest system, which might be
     /// merged into the dest config or might become the dest config
     user_dest: Vec<ignition_config::Config>,
-    /// The coreos-installer config for our own parameters, excluding custom
+    /// The nestos-installer config for our own parameters, excluding custom
     /// configs supplied by the user
     installer: Option<InstallConfig>,
     /// Have the installer copy network configs, if we are running it
@@ -145,11 +145,11 @@ impl LiveInitrd {
     }
 
     pub fn dest_ignition(&mut self, path: &str) -> Result<()> {
-        let data = read(path).with_context(|| format!("reading {}", path))?;
+        let data = read(path).with_context(|| format!("reading {path}"))?;
         let (config, warnings) = ignition_config::Config::parse_slice(&data)
-            .with_context(|| format!("parsing Ignition config {}", path))?;
+            .with_context(|| format!("parsing Ignition config {path}"))?;
         for warning in warnings {
-            eprintln!("Warning parsing {}: {}", path, warning);
+            eprintln!("Warning parsing {path}: {warning}");
         }
         self.user_dest.push(config);
         Ok(())
@@ -191,9 +191,9 @@ impl LiveInitrd {
         if !self.features.live_initrd_network {
             bail!("This OS image does not support customizing network settings.");
         }
-        let data = read(path).with_context(|| format!("reading {}", path))?;
+        let data = read(path).with_context(|| format!("reading {path}"))?;
         let name = filename(path)?;
-        let path = format!("{}/{}", INITRD_NETWORK_DIR, name);
+        let path = format!("{INITRD_NETWORK_DIR}/{name}");
         if self.initrd.get(&path).is_some() {
             bail!("config already specifies keyfile {}", name);
         }
@@ -232,8 +232,10 @@ impl LiveInitrd {
     }
 
     pub fn ignition_ca(&mut self, path: &str) -> Result<()> {
-        let data = read(path).with_context(|| format!("reading {}", path))?;
-        self.live.add_ca(&data)?;
+        let data = read(path).with_context(|| format!("reading {path}"))?;
+        self.live
+            .get_or_insert_with(Default::default)
+            .add_ca(&data)?;
         self.dest_ca.push(data);
         Ok(())
     }
@@ -264,15 +266,12 @@ impl LiveInitrd {
         deps: &str,
         install_target: &str,
     ) -> Result<()> {
-        let data = read(path).with_context(|| format!("reading {}", path))?;
+        let data = read(path).with_context(|| format!("reading {path}"))?;
         let name = filename(path)?;
-        self.live.add_file(
-            format!("/usr/local/bin/{}-install-{}", typ, name),
-            &data,
-            0o700,
-        )?;
-        self.live.add_unit(
-            format!("{}-install-{}.service", typ, name),
+        let live = self.live.get_or_insert_with(Default::default);
+        live.add_file(format!("/usr/local/bin/{typ}-install-{name}"), &data, 0o700)?;
+        live.add_unit(
+            format!("{typ}-install-{name}.service"),
             format!(
                 "# Generated by nestos-installer {{iso|pxe}} customize
 
@@ -301,10 +300,10 @@ RequiredBy={install_target}",
     }
 
     pub fn installer_config(&mut self, path: &str) -> Result<()> {
-        let data = read(path).with_context(|| format!("reading {}", path))?;
+        let data = read(path).with_context(|| format!("reading {path}"))?;
         // we don't validate but at least we parse
         serde_yaml::from_slice::<InstallConfig>(&data)
-            .with_context(|| format!("parsing installer config {}", path))?;
+            .with_context(|| format!("parsing installer config {path}"))?;
         self.installer_config_bytes(&filename(path)?, &data)
     }
 
@@ -312,7 +311,7 @@ RequiredBy={install_target}",
         if !self.features.installer_config {
             bail!("This OS image does not support customizing installer configuration.");
         }
-        self.live.add_file(
+        self.live.get_or_insert_with(Default::default).add_file(
             format!(
                 "/etc/nestos/installer.d/{:04}-{}",
                 self.installer_serial, filename
@@ -325,16 +324,17 @@ RequiredBy={install_target}",
     }
 
     pub fn live_config(&mut self, path: &str) -> Result<()> {
-        let data = read(path).with_context(|| format!("reading {}", path))?;
+        let data = read(path).with_context(|| format!("reading {path}"))?;
         // we don't validate but at least we parse
         let (config, warnings) = ignition_config::Config::parse_slice(&data)
-            .with_context(|| format!("parsing Ignition config {}", path))?;
+            .with_context(|| format!("parsing Ignition config {path}"))?;
         for warning in warnings {
-            eprintln!("Warning parsing {}: {}", path, warning);
+            eprintln!("Warning parsing {path}: {warning}");
         }
         self.live
+            .get_or_insert_with(Default::default)
             .merge_config(&config)
-            .with_context(|| format!("merging Ignition config {}", path))
+            .with_context(|| format!("merging Ignition config {path}"))
     }
 
     pub fn into_initrd(mut self) -> Result<Initrd> {
@@ -365,7 +365,11 @@ RequiredBy={install_target}",
             let conf = self.installer.get_or_insert_with(Default::default);
             assert!(conf.ignition_file.is_none());
             let dest_path = "/etc/nestos/dest.ign";
-            self.live.add_file(dest_path.into(), &data, 0o600)?;
+            self.live.get_or_insert_with(Default::default).add_file(
+                dest_path.into(),
+                &data,
+                0o600,
+            )?;
             conf.ignition_file = Some(dest_path.into());
         }
 
@@ -373,8 +377,7 @@ RequiredBy={install_target}",
             // The installer will run; apply deferred settings
             if let Some(device) = self.installer.as_ref().and_then(|c| c.dest_device.as_ref()) {
                 eprintln!(
-                    "Boot media will automatically install to {} without confirmation.",
-                    device
+                    "Boot media will automatically install to {device} without confirmation."
                 );
             } else {
                 eprintln!("Boot media will automatically run installer.");
@@ -396,8 +399,11 @@ RequiredBy={install_target}",
             )?;
         }
 
-        // Embed live config in initrd
-        self.initrd.add(INITRD_IGNITION_PATH, self.live.to_bytes()?);
+        // Embed live config, if we have one, in the initrd.  Avoid embedding
+        // an empty config because the ISO disables autologin if it sees one.
+        if let Some(live) = self.live {
+            self.initrd.add(INITRD_IGNITION_PATH, live.to_bytes()?);
+        }
         Ok(self.initrd)
     }
 }
