@@ -12,14 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// change coreos to nestos
-
 use anyhow::{bail, Context, Result};
 use lazy_static::lazy_static;
 use serde::Serialize;
 use std::collections::HashMap;
 use std::fs::{create_dir_all, read, File, OpenOptions};
-use std::io::{self, copy, BufReader, BufWriter, Read, Seek, SeekFrom, Write};
+use std::io::{self, copy, BufReader, BufWriter, Read, Seek, Write};
 use std::path::{Component, Path, PathBuf};
 
 use crate::cmdline::*;
@@ -73,7 +71,7 @@ pub fn iso_remove(config: IsoRemoveConfig) -> Result<()> {
 pub fn iso_ignition_embed(config: IsoIgnitionEmbedConfig) -> Result<()> {
     let ignition = match &config.ignition_file {
         Some(ignition_path) => {
-            read(ignition_path).with_context(|| format!("reading {}", ignition_path))?
+            read(ignition_path).with_context(|| format!("reading {ignition_path}"))?
         }
         None => {
             let mut data = Vec::new();
@@ -166,7 +164,7 @@ pub fn pxe_ignition_wrap(config: PxeIgnitionWrapConfig) -> Result<()> {
 
     let ignition = match &config.ignition_file {
         Some(ignition_path) => {
-            read(ignition_path).with_context(|| format!("reading {}", ignition_path))?
+            read(ignition_path).with_context(|| format!("reading {ignition_path}"))?
         }
         None => {
             let mut data = Vec::new();
@@ -192,7 +190,7 @@ pub fn pxe_ignition_unwrap(config: PxeIgnitionUnwrapConfig) -> Result<()> {
             OpenOptions::new()
                 .read(true)
                 .open(path)
-                .with_context(|| format!("opening {}", path))?,
+                .with_context(|| format!("opening {path}"))?,
         )
     } else {
         Box::new(stdin.lock())
@@ -222,9 +220,9 @@ pub fn pxe_network_wrap(config: PxeNetworkWrapConfig) -> Result<()> {
 
 fn initrd_network_embed(initrd: &mut Initrd, keyfiles: &[String]) -> Result<()> {
     for path in keyfiles {
-        let data = read(path).with_context(|| format!("reading {}", path))?;
+        let data = read(path).with_context(|| format!("reading {path}"))?;
         let name = filename(path)?;
-        let path = format!("{}/{}", INITRD_NETWORK_DIR, name);
+        let path = format!("{INITRD_NETWORK_DIR}/{name}");
         if initrd.get(&path).is_some() {
             bail!("multiple input files named '{}'", name);
         }
@@ -240,7 +238,7 @@ pub fn pxe_network_unwrap(config: PxeNetworkUnwrapConfig) -> Result<()> {
             OpenOptions::new()
                 .read(true)
                 .open(path)
-                .with_context(|| format!("opening {}", path))?,
+                .with_context(|| format!("opening {path}"))?,
         )
     } else {
         Box::new(stdin.lock())
@@ -257,7 +255,7 @@ fn initrd_network_extract(initrd: &Initrd, directory: Option<&String>) -> Result
         bail!("No embedded network settings.");
     }
     if let Some(dir) = directory {
-        create_dir_all(&dir)?;
+        create_dir_all(dir)?;
         for (path, contents) in files {
             let path = Path::new(dir).join(filename(path)?);
             OpenOptions::new()
@@ -318,7 +316,7 @@ pub fn iso_kargs_show(config: IsoKargsShowConfig) -> Result<()> {
     } else {
         iso.kargs()?
     };
-    println!("{}", kargs);
+    println!("{kargs}");
     Ok(())
 }
 
@@ -391,7 +389,7 @@ pub fn pxe_customize(config: PxeCustomizeConfig) -> Result<()> {
         path => {
             let dir = Path::new(path)
                 .parent()
-                .with_context(|| format!("no parent directory of {}", path))?;
+                .with_context(|| format!("no parent directory of {path}"))?;
             let tempfile = tempfile::Builder::new()
                 .prefix(".nestos-installer-temp-")
                 .tempfile_in(dir)
@@ -405,7 +403,7 @@ pub fn pxe_customize(config: PxeCustomizeConfig) -> Result<()> {
         INITRD_LIVE_STAMP_PATH,
         INITRD_FEATURES_PATH,
         INITRD_IGNITION_PATH,
-        &format!("{}/*", INITRD_NETWORK_DIR),
+        &format!("{INITRD_NETWORK_DIR}/*"),
     ])
     .unwrap();
     let base_initrd = match &*config.output {
@@ -434,6 +432,11 @@ pub fn pxe_customize(config: PxeCustomizeConfig) -> Result<()> {
 
     let live = LiveInitrd::from_common(&config.common, features)?;
     let initrd = live.into_initrd()?;
+    if initrd.get(INITRD_IGNITION_PATH).is_some() {
+        eprintln!(
+            "PXE configuration must include kernel arguments:\n\tignition.firstboot ignition.platform.id=metal"
+        );
+    }
 
     // append customizations to output
     let do_write = |writer: &mut dyn Write| -> Result<()> {
@@ -448,9 +451,9 @@ pub fn pxe_customize(config: PxeCustomizeConfig) -> Result<()> {
             let mut tempfile = tempfile.unwrap();
             do_write(tempfile.as_file_mut())?;
             tempfile
-                .persist_noclobber(&path)
+                .persist_noclobber(path)
                 .map_err(|e| e.error)
-                .with_context(|| format!("persisting output file to {}", path))?;
+                .with_context(|| format!("persisting output file to {path}"))?;
             Ok(())
         }
     }
@@ -498,7 +501,7 @@ pub fn dev_show_initrd(config: DevShowInitrdConfig) -> Result<()> {
     set_die_on_sigpipe()?;
     let initrd = read_initrd(&config.input, &config.filter)?;
     for path in initrd.find(&ALL_GLOB).keys() {
-        println!("{}", path);
+        println!("{path}");
     }
     Ok(())
 }
@@ -545,7 +548,7 @@ fn read_initrd(path: &str, filter: &[String]) -> Result<Initrd> {
             OpenOptions::new()
                 .read(true)
                 .open(path)
-                .with_context(|| format!("opening {}", path))?,
+                .with_context(|| format!("opening {path}"))?,
             &filter,
         ),
     }
@@ -576,7 +579,7 @@ pub fn iso_extract_pxe(config: IsoExtractPxeConfig) -> Result<()> {
                     s.push(file.name.to_lowercase());
                     s
                 };
-                let path = Path::new(&config.output_dir).join(&filename);
+                let path = Path::new(&config.output_dir).join(filename);
                 println!("{}", path.display());
                 copy_file_from_iso(&mut iso, &file, &path)?;
             }
@@ -611,7 +614,7 @@ pub fn iso_extract_minimal_iso(config: IsoExtractMinimalIsoConfig) -> Result<()>
     if let Some(path) = &config.output_rootfs {
         let rootfs = full_iso
             .get_path(COREOS_ISO_ROOTFS_IMG)
-            .with_context(|| format!("looking up '{}'", COREOS_ISO_ROOTFS_IMG))?
+            .with_context(|| format!("looking up '{COREOS_ISO_ROOTFS_IMG}'"))?
             .try_into_file()?;
         copy_file_from_iso(&mut full_iso, &rootfs, Path::new(path))?;
     }
@@ -621,9 +624,7 @@ pub fn iso_extract_minimal_iso(config: IsoExtractMinimalIsoConfig) -> Result<()>
         Err(e) if e.is::<iso9660::NotFound>() => {
             bail!("This ISO image does not support extracting a minimal ISO.")
         }
-        Err(e) => {
-            return Err(e).with_context(|| format!("looking up '{}'", COREOS_ISO_MINISO_FILE))
-        }
+        Err(e) => return Err(e).with_context(|| format!("looking up '{COREOS_ISO_MINISO_FILE}'")),
     };
 
     let data = {
@@ -632,7 +633,7 @@ pub fn iso_extract_minimal_iso(config: IsoExtractMinimalIsoConfig) -> Result<()>
     };
     let mut outf = tempfile::Builder::new()
         .prefix(".nestos-installer-temp-")
-        .tempfile_in(&output_dir)
+        .tempfile_in(output_dir)
         .context("creating temporary file")?;
     data.unxzpack(full_iso.as_file()?, &mut outf)
         .context("unpacking miniso")?;
@@ -641,7 +642,7 @@ pub fn iso_extract_minimal_iso(config: IsoExtractMinimalIsoConfig) -> Result<()>
         .context("modifying miniso kernel args")?;
 
     if &config.output == "-" {
-        outf.seek(SeekFrom::Start(0))
+        outf.rewind()
             .context("seeking back to start of miniso tempfile")?;
         copy(&mut outf, &mut io::stdout().lock()).context("writing output")?;
     } else {
@@ -672,9 +673,9 @@ pub fn pack_minimal_iso(config: PackMinimalIsoConfig) -> Result<()> {
             .context("packing miniso")?;
     eprintln!("Matched {} files of {}", matches, minimal_files.len());
 
-    eprintln!("Total bytes skipped: {}", skipped);
-    eprintln!("Total bytes written: {}", written);
-    eprintln!("Total bytes written (compressed): {}", written_compressed);
+    eprintln!("Total bytes skipped: {skipped}");
+    eprintln!("Total bytes written: {written}");
+    eprintln!("Total bytes written (compressed): {written_compressed}");
 
     eprintln!("Verifying that packed image matches digest");
     data.unxzpack(full_iso.as_file()?, std::io::sink())
@@ -682,7 +683,7 @@ pub fn pack_minimal_iso(config: PackMinimalIsoConfig) -> Result<()> {
 
     let miniso_entry = full_iso
         .get_path(COREOS_ISO_MINISO_FILE)
-        .with_context(|| format!("looking up '{}'", COREOS_ISO_MINISO_FILE))?
+        .with_context(|| format!("looking up '{COREOS_ISO_MINISO_FILE}'"))?
         .try_into_file()?;
     let mut w = full_iso.overwrite_file(&miniso_entry)?;
     data.serialize(&mut w).context("writing miniso data file")?;
@@ -729,7 +730,7 @@ fn modify_miniso_kargs(f: &mut File, rootfs_url: Option<&String>) -> Result<()> 
             bail!("forbidden whitespace found in '{}'", url);
         }
         let final_kargs = KargsEditor::new()
-            .append(&[format!("nestos.live.rootfs_url={}", url)])
+            .append(&[format!("nestos.live.rootfs_url={url}")])
             .apply_to(&new_default_kargs)?;
 
         cfg.set_kargs(&final_kargs)?;
@@ -738,7 +739,7 @@ fn modify_miniso_kargs(f: &mut File, rootfs_url: Option<&String>) -> Result<()> 
     // update kargs
     write_live_iso(&cfg, f, None)?;
 
-    // also modify the default kargs because we don't want `coreos-installer iso kargs reset` to
-    // re-add `coreos.liveiso`
+    // also modify the default kargs because we don't want `nestos-installer iso kargs reset` to
+    // re-add `nestos.liveiso`
     set_default_kargs(&mut iso, new_default_kargs)
 }
