@@ -12,11 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// change coreos to nestos
 use anyhow::{bail, Context, Result};
 use nix::unistd::isatty;
 use std::fs::{write, File, OpenOptions};
-use std::io::{self, copy, BufWriter, Seek, SeekFrom, Write};
+use std::io::{self, copy, BufWriter, Seek, Write};
 use std::os::unix::io::AsRawFd;
 use std::path::Path;
 
@@ -35,7 +34,7 @@ pub(super) fn open_live_iso(
     OpenOptions::new()
         .read(true)
         .write(matches!(output_path, Some(None)))
-        .open(&input_path)
+        .open(input_path)
         .with_context(|| format!("opening {}", &input_path))
 }
 
@@ -56,18 +55,18 @@ pub(super) fn write_live_iso(
         Some(output_path) => {
             let output_dir = Path::new(output_path)
                 .parent()
-                .with_context(|| format!("no parent directory of {}", output_path))?;
+                .with_context(|| format!("no parent directory of {output_path}"))?;
             let mut output = tempfile::Builder::new()
                 .prefix(".nestos-installer-temp-")
                 .tempfile_in(output_dir)
                 .context("creating temporary file")?;
-            input.seek(SeekFrom::Start(0)).context("seeking input")?;
+            input.rewind().context("seeking input")?;
             copy(input, output.as_file_mut()).context("copying input to temporary file")?;
             iso.write(output.as_file_mut())?;
             output
-                .persist_noclobber(&output_path)
+                .persist_noclobber(output_path)
                 .map_err(|e| e.error)
-                .with_context(|| format!("persisting output file to {}", output_path))?;
+                .with_context(|| format!("persisting output file to {output_path}"))?;
         }
     }
     Ok(())
@@ -78,7 +77,7 @@ pub(super) fn write_live_iso(
 pub(super) fn write_live_pxe(initrd: &Initrd, output_path: Option<&String>) -> Result<()> {
     let initrd = initrd.to_bytes()?;
     match output_path {
-        Some(path) => write(path, &initrd).with_context(|| format!("writing {}", path)),
+        Some(path) => write(path, &initrd).with_context(|| format!("writing {path}")),
         None => {
             let stdout = io::stdout();
             let mut out = stdout.lock();
@@ -96,7 +95,7 @@ pub(super) fn copy_file_from_iso(
     let mut outf = OpenOptions::new()
         .write(true)
         .create_new(true)
-        .open(&output_path)
+        .open(output_path)
         .with_context(|| format!("opening {}", output_path.display()))?;
     let mut bufw = BufWriter::with_capacity(BUFFER_SIZE, &mut outf);
     copy(&mut iso.read_file(file)?, &mut bufw)?;
@@ -114,7 +113,7 @@ pub(super) fn verify_stdout_not_tty() -> Result<()> {
 pub(super) fn filename(path: &str) -> Result<String> {
     Ok(Path::new(path)
         .file_name()
-        .with_context(|| format!("missing filename in {}", path))?
+        .with_context(|| format!("missing filename in {path}"))?
         // path was originally a string
         .to_string_lossy()
         .into_owned())
