@@ -31,10 +31,10 @@ pub(super) fn fiemap_path(path: &OsStr) -> Result<Vec<Extent>> {
     let file = OpenOptions::new()
         .read(true)
         .open(path)
-        .with_context(|| format!("opening {:?}", path))?;
+        .with_context(|| format!("opening {path:?}"))?;
 
     let fd = file.as_raw_fd();
-    fiemap(fd).with_context(|| format!("mapping {:?}", path))
+    fiemap(fd).with_context(|| format!("mapping {path:?}"))
 }
 
 /// Returns the `Extent`s associated with the given file. Note that the physical offsets are
